@@ -4,21 +4,18 @@ apply plugin: 'com.google.firebase.firebase-perf'
 
 android {
     signingConfigs {
-        config {
-            keyAlias 'key0'
-            keyPassword 'dhc9Xe7XoidlRo77hohh'
-            storeFile file('E:/Nextcloud/AndroidApps/BitCoinPools/keyfile.jks')
-            storePassword 'dhc9Xe7XoidlRo77hohh'
+        release {
+            configSign()
         }
     }
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "javinator9889.bitcoinpools"
         minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 75
-        versionName "1.18.6"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 92
+        versionName "1.19"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
