@@ -1,7 +1,6 @@
 package javinator9889.bitcoinpools.AppUpdaterManager;
 
 import android.os.AsyncTask;
-import android.support.annotation.NonNull;
 import android.util.Log;
 
 import org.json.JSONArray;
@@ -15,6 +14,7 @@ import java.io.Reader;
 import java.net.URL;
 import java.nio.charset.Charset;
 
+import androidx.annotation.NonNull;
 import javinator9889.bitcoinpools.Constants;
 
 /**
